% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx-package.R
\docType{package}
\name{cdx}
\alias{cdx}
\alias{cdx-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/cdx}
\item BugReports: \url{https://gitlab.com/hrbrmstr/cdx/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
