% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx-query.R
\name{cdx_query}
\alias{cdx_query}
\title{Query a CDX index endpoint}
\usage{
cdx_query(cdx_api_endpoint, url, include = c("urlkey", "timestamp", "url",
  "mime", "status", "digest", "length", "offset", "filename"), page = 0L,
  from = NULL, to = NULL, match_type = NULL, limit = NULL,
  sort = NULL)
}
\arguments{
\item{cdx_api_endpoint}{the API endpoint to query. If using the Common Crawl CDX
index (or a CDX server that mimics the CC metadata services), it's a good
idea to use \code{\link[=fetch_collections_index]{fetch_collections_index()}} first and then supply a value from
the \code{cdx_api} column as the endpoint.}

\item{url}{host, url, wildcard to search for. e.g. \code{*.example.com}}

\item{include}{which fields to include in the output. The standard available
fields are usually: \code{urlkey}, \code{timestamp}, \code{url}, \code{mime}, \code{status},
\code{digest}, \code{length}, \code{offset}, \code{filename}.}

\item{page}{page is the current page number, and defaults to 0 if omitted.
If the page exceeds the number of available pages from the page count
query, a 400 error will be returned.}

\item{from, to}{Setting \code{from=<ts>} or \code{to=<ts>} will restrict the results to the
given date/time range (inclusive). Timestamps may be <=14 digits and
will be padded to either lower or upper bound. For example,
\code{from = 2014}, \code{to = 2014} will return results that have a timestamp
between \code{20140101000000} and \code{20141231235959}}

\item{match_type}{Optional. If supplied, one of
\itemize{
\item \code{exact}: default setting, will return captures that match the url exactly
\item \code{prefix}: return captures that begin with a specified path, eg: \code{http://example.com/path/*}
\item \code{host}: return captures which for a begin host (the path segment is ignored if specified)
\item \code{domain}: return captures for the current host and all subdomains, eg. \code{*.example.com}
}

As a shortcut, instead of specifying a separate \code{match_type} parameter,
wildcards may be used in the \code{url}. i.e. \code{url = 'http://example.com/path/*'} is
equivalent to \code{url = 'http://example.com/path/'} with \code{match_type = 'prefix'} and
\code{url = 'example.com'} with \code{match_type = 'domain'} is equivalent to
\code{url = '*.example.com'}.}

\item{limit}{limit the number of index lines returned. Limit must be set to
a positive integer. If no limit is provided, all the matching lines
are returned, which may be slow.}

\item{sort}{Options. If supplied, one of:
\itemize{
\item \code{reverse}: will sort the matching captures in reverse order. It is only
recommended for \code{exact} query as reverse a large match may be very slow.
\item \code{closest}: setting this option also requires setting \code{closest = <ts>}
where \code{<ts>} is a specific timestamp to sort by. This option will only
work correctly for \code{exact} query and is useful for sorting captures based
no time distance from a certain timestamp.
}

Both options may be combined with \code{limit} to return the top N closest, or the last N results.}
}
\description{
Query a CDX index endpoint
}
\examples{
cidx <- fetch_collections_index()
rprj <- cdx_query(cidx$cdx_api[1], "*.r-project.org")
}
