% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-cidx.R
\name{fetch_collections_index}
\alias{fetch_collections_index}
\title{Fetch collections index}
\usage{

  fetch_collections_index(cidx_url = "http://index.commoncrawl.org/collinfo.json")
}
\arguments{
\item{cidx_url}{URL to use for the fetch. Defaults to \url{http://index.commoncrawl.org/collinfo.json}}
}
\value{
data frame
}
\description{
Retrieves a collections index list and returns it as a data frame. Default
is to use the \href{http://index.commoncrawl.org/}{Common Crawl Collections Index}
}
\examples{
fetch_collections_index()
}
