% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdx-package.R
\docType{package}
\name{cdx}
\alias{cdx}
\alias{cdx-package}
\title{Query Web Archive Crawl Indexes ('CDX')}
\description{
Methods are provided to retrieve web archive crawl index ('CDX') metadata
and directly query the 'CDX' 'API' endpoint to retrieve mementos for a
given set of parameters.
}
\details{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/cdx}
\item BugReports: \url{https://gitlab.com/hrbrmstr/cdx/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
